+++ link/src/main/java/com/stripe/android/link/ui/LinkButtonView.kt
@@ -40,11 +40,16 @@ import com.stripe.android.link.LinkPaymentLauncher
 import com.stripe.android.link.R
 import com.stripe.android.link.theme.DefaultLinkTheme
 import com.stripe.android.link.theme.linkColors
+import com.stripe.android.uicore.StripeTheme
 
 private val LinkButtonVerticalPadding = 6.dp
 private val LinkButtonHorizontalPadding = 10.dp
-private val LinkButtonShape = RoundedCornerShape(4.dp)
-private val LinkButtonEmailShape = RoundedCornerShape(3.dp)
+private val LinkButtonShape = RoundedCornerShape(
+    StripeTheme.primaryButtonStyle.shape.cornerRadius.dp
+)
+private val LinkButtonEmailShape = RoundedCornerShape(
+    StripeTheme.primaryButtonStyle.shape.cornerRadius.dp / 2
+)
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 const val LinkButtonTestTag = "LinkButtonTestTag"