+++ paymentsheet/src/androidTest/java/com/stripe/android/paymentsheet/ui/GooglePayButtonTest.kt
@@ -6,6 +6,7 @@ import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import com.google.common.truth.Truth.assertThat
+import com.stripe.android.GooglePayJsonFactory
 import com.stripe.android.paymentsheet.MainActivity
 import org.junit.Rule
 import org.junit.Test
@@ -24,7 +25,8 @@ class GooglePayButtonTest {
             GooglePayButton(
                 state = null,
                 isEnabled = true,
-                isDark = false,
+                allowCreditCards = true,
+                billingAddressParameters = GooglePayJsonFactory.BillingAddressParameters(),
                 onPressed = { didCallOnPressed = true },
                 modifier = Modifier.testTag(testTag),
             )
@@ -48,7 +50,8 @@ class GooglePayButtonTest {
             GooglePayButton(
                 state = null,
                 isEnabled = false,
-                isDark = false,
+                allowCreditCards = true,
+                billingAddressParameters = GooglePayJsonFactory.BillingAddressParameters(),
                 onPressed = { didCallOnPressed = true },
                 modifier = Modifier.testTag(testTag),
             )