+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/PaymentSheetViewModel.kt
@@ -13,6 +13,7 @@ import androidx.lifecycle.viewModelScope
 import androidx.lifecycle.viewmodel.CreationExtras
 import com.stripe.android.CreateIntentCallbackForServerSideConfirmation
 import com.stripe.android.ExperimentalPaymentSheetDecouplingApi
+import com.stripe.android.GooglePayJsonFactory
 import com.stripe.android.IntentConfirmationInterceptor
 import com.stripe.android.PaymentConfiguration
 import com.stripe.android.core.Logger
@@ -192,6 +193,23 @@ internal class PaymentSheetViewModel @Inject internal constructor(
             },
         )
     }
+
+    internal val googlePayAllowCreditCards: Boolean
+        get() = googlePayLauncherConfig?.allowCreditCards ?: false
+
+    internal val googlePayBillingAddressParameters: GooglePayJsonFactory.BillingAddressParameters?
+        get() = googlePayLauncherConfig?.let {
+            GooglePayJsonFactory.BillingAddressParameters(
+                it.billingAddressConfig.isRequired,
+                when (it.billingAddressConfig.format) {
+                    GooglePayPaymentMethodLauncher.BillingAddressConfig.Format.Min ->
+                        GooglePayJsonFactory.BillingAddressParameters.Format.Min
+                    GooglePayPaymentMethodLauncher.BillingAddressConfig.Format.Full ->
+                        GooglePayJsonFactory.BillingAddressParameters.Format.Full
+                },
+                it.billingAddressConfig.isPhoneNumberRequired
+            )
+        }
 
     private var paymentLauncher: StripePaymentLauncher? = null